refactor semantic color definitions
